move left portion of EditTrip page to a separate component
as it's state could be separately managed
